Integrate a version of FND's recipe widget.
whereas his lets you select from a list of bags and drop them on the recipe,
and then sort the recipe, this just adds the already dropped bag to the end
of the recipe and then lets you sort the list.

which is a bit cumbersome, but is the single action that I was thinking of.
we'll want both...