Make the description on the baglist page editable.
We do this by making it a contentEditable when it is clicked, and
horking on a save and cancel button. click actions on each of these
either save the bag (described below) or reset the div to its original
value. when saving: we get the bag, reset the desc field, then put the
bag back.